+++ server/systems/GameSystem.cpp
@@ -77,7 +77,7 @@ namespace RType
                     event.push<uint8_t>("mob_id", cMob->getId());
                     event.push<uint32_t>("x", it->second.x);
                     event.push<uint32_t>("y", it->second.y);
-                    event.push<uint32_t>("time",
+                    event.push<uint64_t>("time",
                                          game->getChrono().getElapsedTime());
                     room->broadcastUDP(event.toBuffer());
                     std::cout << "Mob spawned id: " << cMob->getId() << std::endl; // debug
